+++ lib/EpdFont/IEpdFont.h
@@ -12,7 +12,7 @@ class IEpdFont {
   virtual const EpdGlyph* getGlyph(uint32_t cp) const = 0;
   virtual const EpdFontData* getFontData() const = 0;
   // Optional kerning and ligature hooks. Fonts without shaping data can use defaults.
-  virtual int8_t getKerning(uint32_t leftCp, uint32_t rightCp) const {
+  virtual int getKerning(uint32_t leftCp, uint32_t rightCp) const {
     (void)leftCp;
     (void)rightCp;
     return 0;
